refactor: create separate modules for server and web
